updated the test directory and made sure the python script works with it
