+++ .gitignore
@@ -3,4 +3,4 @@
 # build data
 build/
 # log data
-logs/
+log/
