+++ src/dbdiffchecker/FileHandler.java
@@ -14,21 +14,20 @@ import java.util.Scanner;
 
 /**
  * Deals with all data coming from and going out to files.
- * 
+ *
  * @author Peter Kaufman and Chris Dail
  * @version 6-20-20
  * @since 9-12-17
  */
 public class FileHandler {
-  // static variables
   public static final String logFileName = "activity.log";
   public static final String lastSequelStatementFileName = "lastRun.txt";
   public static final String databaseSnapshotFileName = "dbsnapshot";
-  public static final String logFolder = "logs";
+  public static final String logFolder = "log";
 
   /**
    * Serializes a database with all of its table and view data.
-   * 
+   *
    * @author Peter Kaufman
    * @param database The database to serialize.
    * @param prefix   The prefix of the database snapshot to deserailize. <b>Note:
@@ -41,7 +40,6 @@ public class FileHandler {
         FileOutputStream fileOutput = new FileOutputStream(
             new File(logFolder + File.separator + prefix + "_" + databaseSnapshotFileName));
         ObjectOutputStream outputStream = new ObjectOutputStream(fileOutput)) {
-      // Write object to file
       outputStream.writeObject(database);
       outputStream.close();
       fileOutput.close();
@@ -53,7 +51,7 @@ public class FileHandler {
 
   /**
    * Takes SQL statements and writes them to the last run file.
-   * 
+   *
    * @author Peter Kaufman
    * @param sequelStatements Statements to be logged.
    * @throws DatabaseDifferenceCheckerException Error writing the statements to
@@ -73,7 +71,7 @@ public class FileHandler {
 
   /**
    * Takes a String and writes it to the log file.
-   * 
+   *
    * @author Peter Kaufman
    * @param data The data to be written to the log file.
    * @throws DatabaseDifferenceCheckerException Error writing the data to the log
@@ -90,7 +88,7 @@ public class FileHandler {
 
   /**
    * Deserializes a database file.
-   * 
+   *
    * @author Peter Kaufman
    * @param prefix The prefix of the database snapshot to deserailize. <b>Note: it
    *               is the name of the database implimentation of the database</b>
@@ -105,7 +103,6 @@ public class FileHandler {
         FileInputStream fileInput = new FileInputStream(
             new File(logFolder + File.separator + prefix + "_" + databaseSnapshotFileName));
         ObjectInputStream inputStream = new ObjectInputStream(fileInput)) {
-      // Read object
       database = (Database) inputStream.readObject();
     } catch (Exception cause) {
       throw new DatabaseDifferenceCheckerException("There was an error when trying to get the database snapshot.",
@@ -117,7 +114,7 @@ public class FileHandler {
 
   /**
    * Returns the contents of the specified file.
-   * 
+   *
    * @author Peter Kaufman
    * @param file The name of the file to be read from.
    * @return Data that was stored in the file that was read from.
@@ -139,7 +136,7 @@ public class FileHandler {
 
   /**
    * Takes a file path and determines whether the file exists or not.
-   * 
+   *
    * @author Chris Dail
    * @param file The file path to the file.
    * @return Whether the file exists or not.
