+++ src/dbdiffchecker/StopWatch.java
@@ -8,13 +8,12 @@ import java.time.Instant;
  * start and end method calls.
  * 
  * @author Jonas_Hess and Peter Kaufman
- * @version 5-23-19
+ * @version 6-20-20
  * @since 10-26-17
  * @see <a href=
  *      "https://stackoverflow.com/questions/8255738/is-there-a-stopwatch-in-java">https://stackoverflow.com/questions/8255738/is-there-a-stopwatch-in-java</a>
  */
 public class StopWatch {
-  // Instance Variables
   private Instant startTime;
   private Instant endTime;
   private Duration duration;
