+++ src/dbdiffchecker/sql/Column.java
@@ -2,7 +2,7 @@ package dbdiffchecker.sql;
 
 /**
  * Resembles an SQL column and contains column info.
- * 
+ *
  * @author Peter Kaufman
  * @version 5-30-19
  * @since 9-10-17
@@ -12,7 +12,7 @@ public class Column extends Schema {
 
   /**
    * Sets the name and details of the column.
-   * 
+   *
    * @author Peter Kaufman
    * @param name    the name of the column
    * @param details the info of the column
@@ -32,7 +32,7 @@ public class Column extends Schema {
 
   /**
    * Returns the column's details.
-   * 
+   *
    * @author Peter Kaufman
    * @return Details about the column.
    */
