+++ src/dbdiffchecker/sql/SQLiteTable.java
@@ -6,19 +6,18 @@ import java.util.HashMap;
 /**
  * Resembles a table in SQLite and contains info about the table's columns and
  * indices.
- * 
+ *
  * @author Peter Kaufman
  * @version 5-31-19
  * @since 5-11-19
  */
 public class SQLiteTable extends Table {
-  // Instance variables
   private boolean stopCompare = false;
   private int foreignKeyCount = 0;
 
   /**
    * Sets the name and create statement of the table.
-   * 
+   *
    * @author Peter Kaufman
    * @param name   The name of the table.
    * @param create The create statement of the table which will be used to create
@@ -49,7 +48,7 @@ public class SQLiteTable extends Table {
       return sql;
     }
     sql2 += dropIndices(this.indices, t1.getIndices());
-    // if a foreign key is to be dripped, recreate the table
+    // if a foreign key is to be dropped, recreate the table
     if (stopCompare) {
       sql.addAll(recreateTable(t1.getColumns()));
       return sql;
@@ -83,29 +82,22 @@ public class SQLiteTable extends Table {
     String[] parts, columns;
     ArrayList<String> bodySections = new ArrayList<>();
     String indexIndicator = ".*([K|k][E|e][Y|y])(\\s)*(\\().*";
-    String name = "";
-    String drop = "";
-    String details = "";
-    String create = "";
-    String body;
+    String name = "", drop = "", details = "", create = "", body;
     int nameEnd = 0;
     create = createStatement.substring(createStatement.indexOf("(") + 1).trim();
     create = create.trim();
     if (create.endsWith(";")) {
-      create = create.substring(0, create.indexOf(";", create.length() - 6)); // remove last character of create
-                                                                              // statement
+      create = create.substring(0, create.indexOf(";", create.length() - 6));
     }
-    // separate the main create statement from other add ons
+    // separate the main create statement from other add-ons
     parts = create.split(";");
     body = parts[0];
-    // remove unneeded characters
     body = body.trim();
     if (body.endsWith(");")) {
       body = body.substring(0, body.length() - 2);
     } else if (body.endsWith(")")) {
       body = body.substring(0, body.length() - 1);
     }
-    // split the body into parts
     int comma, startParen, endParen;
     while (body.contains(",")) {
       comma = body.indexOf(",");
@@ -166,10 +158,7 @@ public class SQLiteTable extends Table {
   @Override
   protected String dropCols(HashMap<String, Column> cols1, HashMap<String, Column> cols2) {
     String sql = "";
-    Column col = null;
-    // check for columns to drop
     for (String columnName : cols2.keySet()) {
-      col = cols2.get(columnName);
       if (!cols1.containsKey(columnName)) {
         stopCompare = true;
         return sql;
@@ -193,8 +182,8 @@ public class SQLiteTable extends Table {
         this.count++;
       } else {
         col2 = cols2.get(columnName);
-        if (col.getName().equals(col2.getName())) { // columns have the same name
-          if (!col.getDetails().equals(col2.getDetails())) { // column details are different
+        if (col.getName().equals(col2.getName())) {
+          if (!col.getDetails().equals(col2.getDetails())) {
             stopCompare = true;
             return sql;
           }
@@ -207,9 +196,7 @@ public class SQLiteTable extends Table {
   @Override
   protected String dropIndices(HashMap<String, Index> dev, HashMap<String, Index> live) {
     String sql = "";
-    // check for indices to remove
     for (String indexName : live.keySet()) {
-      // if the index does not exist in the dev database then drop it
       if (!dev.containsKey(indexName)) {
         if (live.get(indexName).getCreateStatement().contains("FOREIGN KEY")) {
           stopCompare = true;
@@ -229,9 +216,7 @@ public class SQLiteTable extends Table {
   protected String otherIndices(HashMap<String, Index> dev, HashMap<String, Index> live) {
     String sql = "";
     Index indices1 = null;
-    // check for missing indices
     for (String indexName : dev.keySet()) {
-      // if the index exists in both databases or only in the dev database then add it
       indices1 = dev.get(indexName);
       if (live.containsKey(indexName)) {
         if (!indices1.equals(live.get(indexName))) {
@@ -264,7 +249,7 @@ public class SQLiteTable extends Table {
    * Recreates the structure of the table using the columns provided to copy over
    * old data into the new table based on common columns between the development
    * and live tables.
-   * 
+   *
    * @param live A list of columns and their definitions which helps the transfer
    *             of data for common collumns.
    * @return The SQL statements needed to recreate the development table.
@@ -282,7 +267,6 @@ public class SQLiteTable extends Table {
       // there are columns in common so the table needs to be renamed,
       // have its data copied into a new table of the same name, and then be deleted
       commonColumns = commonColumns.substring(0, commonColumns.length() - 1);
-      // add the appropriate sql
       sql.add("ALTER TABLE " + this.name + " RENAME TO temp_table;");
       if (!doExtraWork) {
         sql.add(this.createStatement);
