+++ src/test/ColumnTest.java
@@ -1,10 +1,12 @@
+package test;
+
 import static org.junit.Assert.assertEquals;
 import org.junit.Test;
 import dbdiffchecker.sql.Column;
 
 /**
  * A unit test that makes sure that the Column object works as intended.
- * 
+ *
  * @author Peter Kaufman
  * @version 5-24-19
  * @since 5-10-19
@@ -17,7 +19,7 @@ public class ColumnTest {
   /**
    * Tests whether the get statements inside of the Column object work as
    * intended.
-   * 
+   *
    * @author Peter Kaufman
    */
   public void testGetStatements() {
