+++ src/test/ViewTest.java
@@ -1,10 +1,12 @@
+package test;
+
 import static org.junit.Assert.assertEquals;
 import org.junit.Test;
 import dbdiffchecker.sql.View;
 
 /**
  * A unit test that makes sure that the View object works as intended.
- * 
+ *
  * @author Peter Kaufman
  * @version 5-24-19
  * @since 5-10-19
@@ -16,7 +18,7 @@ public class ViewTest {
   @Test
   /**
    * Tests whether the get statements inside of the View object work as intended.
-   * 
+   *
    * @author Peter Kaufman
    */
   public void testGetStatements() {
